PUB-1807 - Updated controller, language and tests
