+++ src/main/resources/locales/en/update-user.json
@@ -4,5 +4,5 @@
     "subHeading": "This users current role is ROLE",
     "backButton": "Back",
     "errorTitle": "There is a problem",
-    "errorMessage": "You are unable to update the role for your own user"
+    "errorMessage": "You are unable to update the role for the same user you are logged in as"
 }
